requirements.txt
python-dotenv==1.2.2
requests==2.32.5
openai==2.26.0


Flask==3.0.3